save work

+++ code.h
@@ -70,12 +70,12 @@ public:
     // given 3D point located within the unit cube [0,1].
     unsigned int morton3D(const Point &point)
     {
-        double x = point.x;
-        double y = point.y;
-        double z = point.z;
-        x = std::min(std::max(x * 1024.0, 0.0), 1023.0);
-        y = std::min(std::max(y * 1024.0, 0.0), 1023.0);
-        z = std::min(std::max(z * 1024.0, 0.0), 1023.0);
+        float x = point.x;
+        float y = point.y;
+        float z = point.z;
+        x = std::min(std::max(x * 1024.0f, 0.0f), 1023.0f);
+        y = std::min(std::max(y * 1024.0f, 0.0f), 1023.0f);
+        z = std::min(std::max(z * 1024.0f, 0.0f), 1023.0f);
         unsigned int xx = expandBits((unsigned int)x);
         unsigned int yy = expandBits((unsigned int)y);
         unsigned int zz = expandBits((unsigned int)z);
